Rename types.rs to odbc_types.rs to fix clippy module_inception lint
The `types/types.rs` submodule triggers clippy's `module_inception` lint
because the module shares its name with its parent directory. Renaming to
`odbc_types.rs` resolves the CI failure.

Made-with: Cursor
